scripts/safety/requirements.txt
--index-url https://pypi.tuna.tsinghua.edu.cn/simple/
--find-links https://download.pytorch.org/whl/torch_stable.html

accelerate==0.34.2
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anthropic==0.34.2
antlr4-python3-runtime==4.9.3
anyio==4.5.0
asttokens==2.4.1
async-timeout==4.0.3
attrdict==2.0.1
attrs==24.2.0
beautifulsoup4==4.12.3
braceexpand==0.1.7
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.6
dashscope==1.20.10
datasets==3.0.0
decord==0.6.0
diffusers==0.27.1
dill==0.3.8
distro==1.9.0
einops==0.8.0
exceptiongroup==1.2.2
executing==2.1.0
filelock==3.16.1
# flash-attn==2.6.3
frozenlist==1.4.1
fsspec==2024.6.1
ftfy==6.2.3
google==3.0.0
google-ai-generativelanguage==0.6.9
google-api-core==2.19.2
google-api-python-client==2.146.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.1
googleapis-common-protos==1.65.0
grpcio==1.66.1
grpcio-status==1.66.1
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.2
huggingface-hub==0.25.0
icecream==2.1.3
idna==3.10
importlib_metadata==8.5.0
iopath==0.1.10
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
natsort==8.4.0
networkx==3.2.1
numpy==2.0.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
open_clip_torch==2.26.1
openai==1.46.0
opencv-contrib-python==4.10.0.84
packaging==24.1
pandas==2.2.2
peft==0.4.0
pillow==10.4.0
portalocker==2.10.1
proto-plus==1.24.0
protobuf==5.28.2
psutil==6.0.0
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
rsa==4.9
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.13.1
sentencepiece==0.2.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
sympy==1.13.3
threadpoolctl==3.5.0
timm==1.0.9
tokenizers==0.15.2
torch==2.3.0+cu121
# torch==2.4.1
torchvision==0.18.0+cu121
# torchvision==0.19.1
tqdm==4.66.5
transformers==4.36.0
triton==2.3.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.3
wcwidth==0.2.13
webdataset==0.2.100
websocket-client==1.8.0
xxhash==3.5.0
yarl==1.11.1
zipp==3.20.2
